erp-wagonlits/requirements.txt
flask>=2.3.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dateutil>=2.8.2
